requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# AI/LLM
openai==1.12.0
pydantic==2.6.0

# Data fetching
httpx==0.26.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
python-dateutil==2.8.2
markdown==3.5.2

